Package entry point for module execution
